# docs/openapi/v3.yaml
openapi: 3.1.0
info:
  title: TRAE API
  description: TRAE OpenAPI 3.1 Specification
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.trae.example.com
    description: Production server
components:
  schemas:
    Problem:
      description: RFC 7807 Problem Details
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    DownloadRequest:
      type: object
      required:
        - url
        - shelfId
      properties:
        url:
          type: string
          format: uri
          description: URL of the resource to download
        shelfId:
          type: string
          format: uuid
          description: Target shelf ID
        filename:
          type: string
          description: Custom filename
        metadata:
          type: object
          additionalProperties: true
    Task:
      type: object
      required:
        - id
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SortParam:
      name: sort
      in: query
      description: Sort field and direction
      required: false
      schema:
        type: string
        default: createdAt:desc
    SearchParam:
      name: search
      in: query
      description: Search query string
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication using JWT token
security:
  - BearerAuth: []
paths:
...
  /api/v1/downloads:
    get:
      summary: List all downloads
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: List of downloads
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  total:
                    type: integer
                    minimum: 0
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create download task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
      responses:
        '202':
          description: Download task accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/downloads/{id}:
    get:
      summary: Get download task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Download task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Cancel download task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Download task cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/downloads/{id}/events:
    get:
      summary: Get download task events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: List of download events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - type
                        - message
                        - createdAt
                      properties:
                        id:
                          type: string
                          format: uuid
                        type:
                          type: string
                          enum: [started, progress, completed, failed, cancelled]
                        message:
                          type: string
                        details:
                          type: object
                          additionalProperties: true
                        createdAt:
                          type: string
                          format: date-time
                  total:
                    type: integer
                    minimum: 0
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/plugins:
    get:
      summary: List all plugins
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of plugins
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - name
                        - version
                        - createdAt
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                        version:
                          type: string
                          format: semver
                        description:
                          type: string
                        author:
                          type: string
                        status:
                          type: string
                          enum: [installed, enabled, disabled]
                        capabilities:
                          type: array
                          items:
                            type: string
                        settings:
                          type: object
                          additionalProperties: true
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  total:
                    type: integer
                    minimum: 0
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Install plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                version:
                  type: string
                  format: semver
      responses:
        '202':
          description: Plugin installation initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/plugins/{id}:
    get:
      summary: Get plugin by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plugin details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  version:
                    type: string
                    format: semver
                  description:
                    type: string
                  author:
                    type: string
                  status:
                    type: string
                    enum: [installed, enabled, disabled]
                  capabilities:
                    type: array
                    items:
                      type: string
                  settings:
                    type: object
                    additionalProperties: true
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update plugin settings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                settings:
                  type: object
                  additionalProperties: true
                status:
                  type: string
                  enum: [enabled, disabled]
      responses:
        '200':
          description: Plugin updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  version:
                    type: string
                    format: semver
                  status:
                    type: string
                    enum: [installed, enabled, disabled]
                  settings:
                    type: object
                    additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Uninstall plugin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Plugin uninstalled successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...